week_2/requirements.txt
agno
python-dotenv
Groq
duckduckgo-search
yfinance
streamlit